Move version check into its own file in the CCU assembly; Split asmdef patch into PolyToolkit and SpatialTracking; Update readme to reflect current setup process; Add stub class with serialized fields to ensure EditorVR default references survive package import
