--- Patches/Asmdef_Dependencies_Pre_2019_1.unitypackage
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:92141740e6a868417b9d76f3cef3dd6acae8d91c8525a7d98cdbbadefe751a9c
-size 1462
