+++ Patches/SpatialTracking_2018.4.unitypackage
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2e3581333ae4af126957d8a0e67c5931d819a3ca176ec387094711973ff14398
+size 627
