+++ Patches/SpatialTracking_2018.4.unitypackage.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6abfefbdb52a5274ba0e5293dfe29473
+guid: 06d3631214f49e044a2e0f9fcf6dc5b6
 DefaultImporter:
   externalObjects: {}
   userData:
