+++ Scripts/Utilities/Editor/EXR-Utilities-Editor.asmdef
@@ -1,10 +1,14 @@
 {
     "name": "EXR-Utilities-Editor",
-    "references": [
-        "EXR"
-    ],
+    "references": [],
+    "optionalUnityReferences": [],
     "includePlatforms": [
         "Editor"
     ],
-    "excludePlatforms": []
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": []
 }
